requirements.txt
pandas
protobuf
grpcio
grpcio-tools
numpy
xlrd
sklearn
filetype
Pillow
matplotlib
seaborn
flask
connexion[swagger-ui]